refactor: move shared Event/Project components to lib
