+++ src/lib/components/EventCard.svelte
