+++ src/lib/components/EventsPanel.svelte
